backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
alembic
pyserial
requests    
pyjwt
passlib
bcrypt